# docs/swagger.yaml
definitions:
  common.Response:
    properties:
      description:
        type: string
      internal_code:
        type: integer
    type: object
  internal_account_model.BusinessDataLogic:
    properties:
      economic_activity:
        type: string
      inn:
        type: string
      name:
        type: string
      website:
        type: string
    type: object
  internal_account_model.PersonalDataLogic:
    properties:
      email:
        type: string
      full_name:
        type: string
      geography:
        type: string
      job_position:
        type: string
    type: object
...
  model.ChangeConstantUnitLogic:
    properties:
      category:
        type: string
      name:
        type: string
      value: {}
    type: object
  model.ChangeConstantsRequest:
    properties:
      elements:
        items:
          $ref: '#/definitions/model.ChangeConstantUnitLogic'
        type: array
    type: object
...
  model.GetCommonInfoResponse:
    properties:
      business_data:
        $ref: '#/definitions/internal_account_model.BusinessDataLogic'
      personal_data:
        $ref: '#/definitions/internal_account_model.PersonalDataLogic'
    type: object
  model.GetConstantsResponse:
    properties:
      county_prices:
        items:
          properties:
            county_id:
              type: integer
            county_name:
              type: string
            county_price:
              type: number
          type: object
        type: array
      machine_prices:
        items:
          properties:
            machine_id:
              type: integer
            machine_name:
              type: string
            machine_price:
              type: integer
          type: object
        type: array
      mean_salaries:
        items:
          properties:
            industry_id:
              type: integer
            industry_name:
              type: string
            salary:
              type: integer
          type: object
        type: array
      other_needs:
        items:
          properties:
            need_coeff:
              type: number
            need_id:
              type: integer
            need_name:
              type: string
          type: object
        type: array
      patent_prices:
        items:
          properties:
            patent_id:
              type: integer
            patent_name:
              type: string
            patent_price:
              type: integer
          type: object
        type: array
    type: object
  model.GetResultsByUserResponse:
    properties:
      results:
        items:
          properties:
            name:
              type: string
            report_id:
              type: string
            summary:
              type: integer
            time_stamp:
              type: string
          type: object
        type: array
    type: object
...
  model.UpdateUserInfoRequest:
    properties:
      business_data:
        $ref: '#/definitions/internal_account_model.BusinessDataLogic'
      personal_data:
        $ref: '#/definitions/internal_account_model.PersonalDataLogic'
    type: object
...
host: gate.gb.ldt2023.infantem.tech
info:
  contact:
    email: KlenoviySir@yandex.ru
    name: Docs developer
    url: https://t.me/KlenoviySirop
  description: JWT token in authorization bearer
  title: Core backend app for Leaders of Digital Transformation
  version: 1.0.0
paths:
  /account/constant:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetConstantsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.Response'
      summary: Get all constants for formula
      tags:
      - Account
      - Admin
    patch:
      description: Endpoint to change constants
      parameters:
      - description: Change constants body
        in: body
        name: updated_data
        required: true
        schema:
          $ref: '#/definitions/model.ChangeConstantsRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.Response'
      summary: Change constants for formula
      tags:
      - Account
      - Admin
  /account/info:
    get:
      description: Endpoint to get information about user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetCommonInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.Response'
      summary: Get information about user
      tags:
      - Account
    put:
      description: Endpoint to get information about user
      parameters:
      - description: Updated information about user
        in: body
        name: updated_data
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserInfoRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetCommonInfoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.Response'
      summary: Update user's information
      tags:
      - Account
  /account/results:
    get:
      description: Endpoint to get all information by user (temp w/o pagination)
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetResultsByUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/common.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/common.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/common.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/common.Response'
      summary: Get all results for account
      tags:
      - Account
...
schemes:
- https
swagger: "2.0"